backend/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
gunicorn==20.1.0
pytest==7.3.2
PyJWT==2.8.0
